remove attempts at thread safety and mark buffer sequence as unsafe
