+++ sdk/core/System.ClientModel/src/ModelReaderWriter/ModelWriterOfT.cs
@@ -26,9 +26,9 @@ internal partial class ModelWriter<T>
         _options = options;
     }
 
-    public BufferSequence.Reader GetSequenceBufferReader()
+    public UnsafeBufferSequence.Reader GetSequenceBufferReader()
     {
-        using BufferSequence sequenceWriter = new BufferSequence();
+        using UnsafeBufferSequence sequenceWriter = new UnsafeBufferSequence();
         using var jsonWriter = new Utf8JsonWriter(sequenceWriter);
         _model.Write(jsonWriter, _options);
         jsonWriter.Flush();
