+++ sdk/core/System.ClientModel/tests/internal.perf/ModelReaderWriter/JsonModelBenchmark.cs
@@ -73,10 +73,9 @@ namespace System.ClientModel.Tests.Internal.Perf
 
         [Benchmark]
         [BenchmarkCategory("ModelReaderWriter")]
-        public bool Write_ModelWriter()
+        public void Write_ModelWriter()
         {
             using var reader = new ModelWriter(_model, _options).GetSequenceBufferReader();
-            return reader.TryComputeLength(out var length);
         }
 
         [Benchmark]
