+++ sdk/core/System.ClientModel/tests/internal/ModelReaderWriter/UnsafeBufferSequenceTests.cs
@@ -7,21 +7,20 @@ using NUnit.Framework;
 
 namespace System.ClientModel.Tests.Internal.ModelReaderWriterTests
 {
-    public class BufferSequenceTests
+    public class UnsafeBufferSequenceTests
     {
-        private static readonly FieldInfo _buffersField = typeof(BufferSequence).GetField("_buffers", BindingFlags.NonPublic | BindingFlags.Instance)!;
+        private static readonly FieldInfo _buffersField = typeof(UnsafeBufferSequence).GetField("_buffers", BindingFlags.NonPublic | BindingFlags.Instance)!;
 
         [Test]
         public void ZeroOutAfterGettingReader()
         {
-            using BufferSequence writer = BufferSequenceHelper.SetUpBufferBuilder();
+            using UnsafeBufferSequence writer = UnsafeBufferSequenceHelper.SetUpBufferBuilder();
             ClientModel.Internal.BufferSegment[] buffers = (ClientModel.Internal.BufferSegment[])_buffersField.GetValue(writer)!;
             Assert.IsNotNull(buffers);
             Assert.Greater(buffers.Length, 0);
 
-            using BufferSequence.Reader reader = writer.ExtractSequenceBufferReader();
-            Assert.IsTrue(reader.TryComputeLength(out long length));
-            Assert.AreEqual(100000, length);
+            using UnsafeBufferSequence.Reader reader = writer.ExtractSequenceBufferReader();
+            Assert.AreEqual(100000, reader.Length);
 
             buffers = (ClientModel.Internal.BufferSegment[])_buffersField.GetValue(writer)!;
             Assert.IsNotNull(buffers);
@@ -31,17 +30,16 @@ namespace System.ClientModel.Tests.Internal.ModelReaderWriterTests
         [Test]
         public void CanContinueToWriteAfterZeroOut()
         {
-            using BufferSequence writer = BufferSequenceHelper.SetUpBufferBuilder();
-            using BufferSequence.Reader reader = writer.ExtractSequenceBufferReader();
+            using UnsafeBufferSequence writer = UnsafeBufferSequenceHelper.SetUpBufferBuilder();
+            using UnsafeBufferSequence.Reader reader = writer.ExtractSequenceBufferReader();
 
-            BufferSequenceHelper.WriteBytes(writer, 0xFF, 110000, 15000);
+            UnsafeBufferSequenceHelper.WriteBytes(writer, 0xFF, 110000, 15000);
             ClientModel.Internal.BufferSegment[] buffers = (ClientModel.Internal.BufferSegment[])_buffersField.GetValue(writer)!;
             Assert.IsNotNull(buffers);
             Assert.Greater(buffers.Length, 0);
 
-            using BufferSequence.Reader reader2 = writer.ExtractSequenceBufferReader();
-            Assert.IsTrue(reader2.TryComputeLength(out long length));
-            Assert.AreEqual(110000, length);
+            using UnsafeBufferSequence.Reader reader2 = writer.ExtractSequenceBufferReader();
+            Assert.AreEqual(110000, reader2.Length);
 
             buffers = (ClientModel.Internal.BufferSegment[])_buffersField.GetValue(writer)!;
             Assert.IsNotNull(buffers);
@@ -51,7 +49,7 @@ namespace System.ClientModel.Tests.Internal.ModelReaderWriterTests
         [Test]
         public void DisposeZerosOutBuffers()
         {
-            using BufferSequence writer = BufferSequenceHelper.SetUpBufferBuilder();
+            using UnsafeBufferSequence writer = UnsafeBufferSequenceHelper.SetUpBufferBuilder();
             ClientModel.Internal.BufferSegment[] buffers = (ClientModel.Internal.BufferSegment[])_buffersField.GetValue(writer)!;
             Assert.IsNotNull(buffers);
             Assert.Greater(buffers.Length, 0);
@@ -65,17 +63,16 @@ namespace System.ClientModel.Tests.Internal.ModelReaderWriterTests
         [Test]
         public void CanContinueToWriteAfterDispose()
         {
-            BufferSequence writer = BufferSequenceHelper.SetUpBufferBuilder();
+            UnsafeBufferSequence writer = UnsafeBufferSequenceHelper.SetUpBufferBuilder();
             writer.Dispose();
 
-            BufferSequenceHelper.WriteBytes(writer, 0xFF, 110000, 15000);
+            UnsafeBufferSequenceHelper.WriteBytes(writer, 0xFF, 110000, 15000);
             ClientModel.Internal.BufferSegment[] buffers = (ClientModel.Internal.BufferSegment[])_buffersField.GetValue(writer)!;
             Assert.IsNotNull(buffers);
             Assert.Greater(buffers.Length, 0);
 
-            using BufferSequence.Reader reader = writer.ExtractSequenceBufferReader();
-            Assert.IsTrue(reader.TryComputeLength(out long length));
-            Assert.AreEqual(110000, length);
+            using UnsafeBufferSequence.Reader reader = writer.ExtractSequenceBufferReader();
+            Assert.AreEqual(110000, reader.Length);
 
             buffers = (ClientModel.Internal.BufferSegment[])_buffersField.GetValue(writer)!;
             Assert.IsNotNull(buffers);
